+++ spec/cucumber/rb_support/rb_step_definition_spec.rb
@@ -22,7 +22,7 @@ module Cucumber
       
       it "should allow calling of other steps" do
         dsl.Given /Outside/ do
-          Given "Inside"
+          step "Inside"
         end
         dsl.Given /Inside/ do
           $inside = true
@@ -34,7 +34,7 @@ module Cucumber
 
       it "should allow calling of other steps with inline arg" do
         dsl.Given /Outside/ do
-          Given "Inside", Cucumber::Ast::Table.new([['inside']])
+          step "Inside", Cucumber::Ast::Table.new([['inside']])
         end
         dsl.Given /Inside/ do |table|
           $inside = table.raw[0][0]
@@ -62,7 +62,7 @@ module Cucumber
 
       it "should raise Undefined when inside step is not defined" do
         dsl.Given /Outside/ do
-          Given 'Inside'
+          step 'Inside'
         end
 
         lambda do