+++ History.md
@@ -13,6 +13,10 @@ New format:
     [feature here]
  
 Also see [Gherkin 2.6.0 History](https://github.com/cucumber/gherkin/blob/master/History.md) for info about new `id` and `uri` elements in the JSON.
+## In Git
+
+### Changed features
+* Deprecated i18n methods in World, and added a new #step method to use instead. ([#68](https://github.com/cucumber/cucumber/issues/68) Matt Wynne)
 
 ## [v1.1.0](https://github.com/cucumber/cucumber/compare/v1.0.6...v1.1.0)